requirements.txt
cx-Oracle
dnspython
Flask
Flask-Cors
IPy
kazoo
paramiko
pycryptodome
pyDes
pymongo
PyMySQL
pyOpenSSL
python-masscan
python-nmap
requests
urllib3
uwsgi
Werkzeug
psycopg2

